pyproject.toml
[project]
name = "genai"
version = "0.1.0"
description = "AI-powered text processing and vector search with Pinecone"
authors = [
    {name = "Sahil Sharma", email = "sahil1010sharma@gmail.com"}
]
requires-python = ">=3.12"
dependencies = [
    "pinecone>=7.0.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-pinecone>=0.2.0",
    "langchain-huggingface>=0.3.0",
    "sentence-transformers>=2.0.0",
    "transformers>=4.40.0",
    "torch>=2.0.0,<2.8.0",
    "pinecone-datasets>=0.1.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "openai>=1.0.0",
    "groq>=0.4.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
